+++ next-docs/components/templates/social.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 import {Button} from "@heathmont/moon-core-tw";
-import Bitcasino from "../icons/bitcasino";
-import Instagram from "../icons/instagram";
-import Telegram from "../icons/telegram";
-import Tiktok from "../icons/tiktok";
-import Twitter from "../icons/twitter";
-import Youtube from "../icons/youtube";
+import Bitcasino from "../../pages/templates/icons/bitcasino";
+import Instagram from "../../pages/templates/icons/instagram";
+import Telegram from "../../pages/templates/icons/telegram";
+import Tiktok from "../../pages/templates/icons/tiktok";
+import Twitter from "../../pages/templates/icons/twitter";
+import Youtube from "../../pages/templates/icons/youtube";
 
 interface Props {
   title: string;
@@ -20,7 +20,7 @@ const Social = ({
     <p className="text-[32px] text-center font-semibold p-4 mb-4">
       {title}
     </p>
-    <div className="flex gap-2">
+    <div className="flex gap-2 self-center items-center">
       <Button
         iconOnly
         onClick={() => onClick('telegram')}
@@ -29,7 +29,7 @@ const Social = ({
           backgroundColor: '#27A4DB',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Telegram />
         </div>
       </Button>
@@ -41,7 +41,7 @@ const Social = ({
           backgroundColor: '#FF0000',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Youtube />
         </div>
       </Button>
@@ -53,7 +53,7 @@ const Social = ({
           backgroundColor: '#F7931A',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Bitcasino />
         </div>
       </Button>
@@ -65,7 +65,7 @@ const Social = ({
           backgroundColor: '#1DA1F2',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Twitter />
         </div>
       </Button>
@@ -78,7 +78,7 @@ const Social = ({
             'linear-gradient(49.12deg, #FFDD55 6.62%, #FF5D3F 47.35%, #FF543E 50.06%, #C837AB 93.5%)',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Instagram />
         </div>
       </Button>
@@ -90,7 +90,7 @@ const Social = ({
           backgroundColor: '#000',
         }}
       >
-        <div className="w-[36px] h-[36px] flex justify-center items-center">
+        <div className="w-[28px] h-[28px] flex justify-center items-center">
           <Tiktok />
         </div>
       </Button>